+++ CHANGELOG.md
@@ -7,7 +7,8 @@ You now need the latest version of your Linux distribution or at least Windows 1
 ## Changes compared to v1.1.2.5410
 
 - As for the last two feature updates, this one gets another codename: "Garuka Pula".
-- The feature to export VMs has been added.
+- The feature to import and export VMs has been added.
+- The readme has been updated with the new system requirements.
 
 # EmuGUI v1.1.2.5410 "Sara Angeline" (based on v1.1.1.5409)